separate Selection class from SelectOperation.js
