+++ lib/queryBuilder/QueryBuilder.js
@@ -33,7 +33,7 @@ const FindByIdOperation = require('./operations/FindByIdOperation');
 const FindByIdsOperation = require('./operations/FindByIdsOperation');
 const OnBuildOperation = require('./operations/OnBuildOperation');
 const OnErrorOperation = require('./operations/OnErrorOperation');
-const SelectOperation = require('./operations/SelectOperation');
+const SelectOperation = require('./operations/select/SelectOperation');
 const EagerOperation = require('./operations/eager/EagerOperation');
 const RangeOperation = require('./operations/RangeOperation');
 const FirstOperation = require('./operations/FirstOperation');
@@ -145,9 +145,11 @@ class QueryBuilder extends QueryBuilderBase {
     for (let i = 0, l = arguments.length; i < l; ++i) {
       const name = arguments[i];
       const filter = namedFilters[name];
+
       if (typeof filter !== 'function') {
         throw new Error(`Could not find filter "${name}".`);
       }
+
       filter(this);
     }
