+++ lib/queryBuilder/QueryBuilderBase.js
@@ -2,7 +2,7 @@ const QueryBuilderOperationSupport = require('./QueryBuilderOperationSupport');
 const { isSqlite, isMsSql } = require('../utils/knexUtils');
 
 const KnexOperation = require('./operations/KnexOperation');
-const SelectOperation = require('./operations/SelectOperation');
+const SelectOperation = require('./operations/select/SelectOperation');
 const ReturningOperation = require('./operations/ReturningOperation');
 const WhereCompositeOperation = require('./operations/WhereCompositeOperation');
 const WhereInCompositeOperation = require('./operations/whereInComposite/WhereInCompositeOperation');
