+++ lib/queryBuilder/operations/select/SelectOperation.js
@@ -1,13 +1,12 @@
 const flatten = require('lodash/flatten');
-const ObjectionToKnexConvertingOperation = require('./ObjectionToKnexConvertingOperation');
+const Selection = require('./Selection');
+const ObjectionToKnexConvertingOperation = require('../ObjectionToKnexConvertingOperation');
 
-const ALIAS_REGEX = /\s+as\s+/i;
 const COUNT_REGEX = /count/i;
 
 class SelectOperation extends ObjectionToKnexConvertingOperation {
   constructor(name, opt) {
     super(name, opt);
-
     this.selections = [];
   }
 
@@ -16,31 +15,7 @@ class SelectOperation extends ObjectionToKnexConvertingOperation {
   }
 
   static parseSelection(selection) {
-    let dotIdx, table, column;
-    let alias = null;
-
-    if (typeof selection !== 'string') {
-      return null;
-    }
-
-    if (ALIAS_REGEX.test(selection)) {
-      const parts = selection.split(ALIAS_REGEX);
-
-      selection = parts[0].trim();
-      alias = parts[1].trim();
-    }
-
-    dotIdx = selection.lastIndexOf('.');
-
-    if (dotIdx !== -1) {
-      table = selection.substr(0, dotIdx);
-      column = selection.substr(dotIdx + 1);
-    } else {
-      table = null;
-      column = selection;
-    }
-
-    return new this.Selection(table, column, alias);
+    return this.Selection.create(selection);
   }
 
   onAdd(builder, args) {
@@ -89,29 +64,5 @@ class SelectOperation extends ObjectionToKnexConvertingOperation {
     return null;
   }
 }
-
-class Selection {
-  constructor(table, column, alias) {
-    this.table = table || null;
-    this.column = column || null;
-    this.alias = alias || null;
-  }
-
-  static get SelectAll() {
-    return SELECT_ALL;
-  }
-
-  get name() {
-    return this.alias || this.column;
-  }
-
-  selects(that) {
-    const tablesMatch = that.table === this.table || this.table === null || that.table === null;
-    const colsMatch = this.column === that.column || this.column === '*';
-    return tablesMatch && colsMatch;
-  }
-}
-
-const SELECT_ALL = new Selection(null, '*');
 
 module.exports = SelectOperation;
